restructured parser
