+++ srcs/app/test/config_analyser/ParserTest.cpp
@@ -6,7 +6,6 @@
 #include <parser/Token.hpp>
 #include <parser/Preprocessor.hpp>
 #include <parser/Parser.hpp>
-#include <parser/ParserAPI.hpp>
 #include <parser/ConfigSetters.hpp>
 #include <Config.hpp>
